refactor: share provider-call preparation between Translator and Previewer
Previewer and Translator carried verbatim copies of passage,
same_language?, ensure_supported!, the "cannot detect the source
language" refusal and the SentenceCache construction -- the exact
duplication that let the opaque_elements argument be forgotten in
both places. Extract the shared parts into CallPreparation, included
by both, so the next omission cannot happen twice. Each class keeps
its own Error and does not merge into the other.
