+++ lib/translation_diff.rb
@@ -57,6 +57,7 @@ require "translation_diff/rate_limiters"
 require "translation_diff/redis_rate_limiter"
 require "translation_diff/active_record_rate_limiter"
 require "translation_diff/instrumentation"
+require "translation_diff/call_preparation"
 require "translation_diff/dispatcher"
 require "translation_diff/translator"
 require "translation_diff/preview"
